added ContactEvent hpp file
